Refactor: PostContent에 비즈니스 로직 다 내리기
