+++ src/widgets/post/PostContent.tsx
@@ -1,6 +1,5 @@
-// src/widgets/post/PostContent.tsx
-
 import { useAtom } from 'jotai';
+import { useLocation } from 'react-router-dom';
 import { PostFilters } from '@/features/post/ui/PostsFilter';
 import { PostTable } from '@/features/post/ui/PostTable';
 import { PostPagination } from '@/features/post/ui/PaginationControls';
@@ -14,9 +13,10 @@ import {
 } from '@/features/post/model/atom';
 import { searchQueryAtom } from '@/features/post/model/atom';
 import { usePostHandlers } from '@/entities/post/model/usePostHandlers';
-
+import { useEffect } from 'react';
 import { Post } from '@/entities/post/model/type';
 import { User } from '@/entities/user/model/type';
+import { useUpdateURL } from '@/features/post/model/useUpdateURL';
 
 interface PostContentProps {
   openUserModal: (user: User) => void;
@@ -27,6 +27,7 @@ export const PostContent = ({
   openUserModal,
   openPostDetail,
 }: PostContentProps) => {
+  const location = useLocation();
   const [posts] = useAtom(postsAtom);
   const [total] = useAtom(totalAtom);
   const [skip, setSkip] = useAtom(skipAtom);
@@ -36,19 +37,47 @@ export const PostContent = ({
   const [sortOrder, setSortOrder] = useAtom(sortOrderAtom);
   const [selectedTag, setSelectedTag] = useAtom(selectedTagAtom);
   const [tags] = useAtom(tagsAtom);
+  const { searchPosts, fetchPostsByTag, deletePost, fetchPosts, fetchTags } =
+    usePostHandlers();
+  const { updateURL } = useUpdateURL();
 
-  const { searchPosts, fetchPostsByTag, deletePost } = usePostHandlers();
+  useEffect(() => {
+    fetchTags();
+  }, [fetchTags]);
 
-  const updateURL = () => {
-    const params = new URLSearchParams();
-    if (skip) params.set('skip', skip.toString());
-    if (limit) params.set('limit', limit.toString());
-    if (searchQuery) params.set('search', searchQuery);
-    if (sortBy) params.set('sortBy', sortBy);
-    if (sortOrder) params.set('sortOrder', sortOrder);
-    if (selectedTag) params.set('tag', selectedTag);
-    window.history.replaceState(null, '', `?${params.toString()}`);
-  };
+  useEffect(() => {
+    if (selectedTag) fetchPostsByTag(selectedTag);
+    else fetchPosts();
+
+    updateURL();
+  }, [
+    skip,
+    limit,
+    sortBy,
+    sortOrder,
+    selectedTag,
+    updateURL,
+    fetchPostsByTag,
+    fetchPosts,
+  ]);
+
+  useEffect(() => {
+    const params = new URLSearchParams(location.search);
+    setSkip(parseInt(params.get('skip') || '0'));
+    setLimit(parseInt(params.get('limit') || '10'));
+    setSearchQuery(params.get('search') || '');
+    setSortBy(params.get('sortBy') || '');
+    setSortOrder((params.get('sortOrder') as 'asc' | 'desc') || 'asc');
+    setSelectedTag(params.get('tag') || '');
+  }, [
+    location.search,
+    setLimit,
+    setSkip,
+    setSearchQuery,
+    setSortBy,
+    setSortOrder,
+    setSelectedTag,
+  ]);
 
   return (
     <div className="flex flex-col gap-4">
